Create primecalc2n.py
This calculates a prime number based on the index value put in at the command line as an argument.  If you want to get the 10th prime number, run this from the command line as follows:

python primecalc2n.py 10

It will yield the result of 29 (the 10th prime number)
NOTE: on systems that have that have older python versions (i.e. Python 2.7.x) installed as well as Python 3.x
you would need to precede the script with python3